requirements.txt
notebook 
python-dotenv 
opencv-python 
requests 
pillow 
flask 